services/lfg_api/lfg_bank/requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==4.2.5
django-celery-results==2.5.1
django-extensions==3.2.3
django-rest==0.8.7
djangorestframework==3.14.0
executing==1.2.0
ipython==7.16.0
jedi==0.19.0
kombu==5.3.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
